Some improvements for translations
There are some strings that can't be translated because they aren't translation ready.

In addition some strings got the wrong translation domain.

Some of the sprintf-functions are changed because they include URLs in the translated string. This will cause a new translation if the URL is changed in the futere. I think the change is a better solution so the strings are now also translateable with the "CodeStyling Localization" plugin.