Change to intialization of 'Color Selectors', 'Sortable Columns', 'Add Last Class to Elements', 'Show/Hide linked elements', 'RTE'.
explained:

Big change to JS intialization of elements that modify the DOM when WP Customizer loads up - e.g. Color Selectors, Rich Text Editors.
-------------------------
These intializers modify the DOM and when done on a site with lots of Layers pages, meaning lots of widgets, they can impact performance at load time and cause hangs and crashes - so we've changed strategy: On page load we will initialize all these type of elements in the Controls but not the Widgets. Then rather initialize them as they are needed when our widget is focused or added. We will continue to test the performance results and hope it makes using Layers even quicker.